Attempted to implement file CRC checking but it doesn't work.
Although the algorithm produces the same results as the presumed
hardware specifies, the check fails on the provided OTA firmware
blobs.  These blobs are only 116 KB whereas the hardware has
128 KB of flash.  So perhaps not all of the flash is overwritten
during the OTA firmware upload or a different integrity checking
algorithm is used.  Either way, the code to check the CRCs is
disabled for now.